backend/requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
sqlalchemy==2.0.27
sqlalchemy-serializer==1.4.1
werkzeug==2.3.7
marshmallow==3.20.1
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
gunicorn==21.2.0
requests==2.31.0
flask-cors==4.0.0
